+++ queen/fitness.go
@@ -7,7 +7,7 @@ func (p *phenotype) CalcFitness() float64 {
 	count := 0
 	for i := 0; i < len(p.board); i++ {
 		for j := i + 1; j < len(p.board); j++ {
-			if (j-i)*(j-i) == (p.board[j]-p.board[i])*(p.board[j]-p.board[i]) {
+			if (j-i)*(j-i) == int((p.board[j]-p.board[i])*(p.board[j]-p.board[i])) {
 				count++
 			}
 		}
